Refactored some things out to make them nicer
