+++ lta/crypto.py
@@ -1,0 +1,16 @@
+# crypto.py
+"""Module that provides cryptographic support services."""
+
+import hashlib
+
+# Adapted from: https://stackoverflow.com/a/44873382
+def sha512sum(filename: str) -> str:
+    """Compute the SHA512 hash of the data in the specified file."""
+    h = hashlib.sha512()
+    b = bytearray(128*1024)
+    mv = memoryview(b)
+    with open(filename, 'rb', buffering=0) as f:
+        # Known issue with MyPy: https://github.com/python/typeshed/issues/2166
+        for n in iter(lambda: f.readinto(mv), 0):  # type: ignore
+            h.update(mv[:n])
+    return h.hexdigest()
